Server/requirements.txt
azure_storage==0.37.0
Flask==3.0.3
Flask_Cors==4.0.0
langchain==0.1.16
openai==1.17.0
pymongo==4.6.3
PyPDF2==3.0.1
